beta 1.2.0 minor bug fixed

+++ app/Http/Controllers/Config/ConfigController.php
@@ -5,12 +5,31 @@ namespace App\Http\Controllers\Config;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Models\Config;
+use Illuminate\Support\Facades\Gate;
+
 
 class ConfigController extends Controller
-{
+{ 
+   public function __construct()
+  {
+    $this->middleware('auth')->except(['dashboard']);
+  }
+
     public function create()
     {        
-      return view('config/create');
+      
+      $response=Gate::inspect('only-admin');
+      if($response->allowed())
+      {
+        return view('config/create');
+      }
+
+      else
+      {
+        return back()->with('info',"You are not Authorize ! to create ");
+      }  
+     
+      
     }
     public function store(Request $request)
     {
@@ -51,49 +70,73 @@ class ConfigController extends Controller
 
     public function detail($id)
     {
+      $response=Gate::inspect('only-admin');
+      if($response->allowed())
+      {
         $data=Config::find($id);
         return view('/config/detail',['configs'=>$data]);
+      }
+
+      else
+      {
+        return back()->with('info',"You are not Authorize ! to update ");
+      }
+      
+        
     }
 
 
     public function update(Request $request,$id)
     {
-      $validate=validator($request->all(),
-      [
-
-        'id'=>'required',
-        'configkey'=>'required',
-        'name'=>'required',
-        'description'=>'required',
-        'isActive'=>'required',
-        'remark'=>'required',
+    //  dd($request);
+        $validate=validator($request->all(),
+        [
+  
+          'id'=>'required',
+          'configkey'=>'required',
+          'name'=>'required',
+          'description'=>'required',
+          'isActive'=>'required',
+          'remark'=>'required',
+       
+        ]);
+        if($validate->fails())
+        {
+            return back()->withErrors($validate);
+        }
+        else
+        {
+          $configs=Config::find($id);
+          $configs->id=$request->id;
+          $configs->configkey=request()->configkey;
+          $configs->name=request()->name;
+          $configs->description=request()->description;
+          $configs->isActive=request()->isActive;
+          $configs->remark=request()->remark;
+          $configs->update();
+          return redirect('config/dashboard')
+          ->with('info','Successfully Update !');
+        }
      
-      ]);
-      if($validate->fails())
-      {
-          return back()->withErrors($validate);
-      }
-      else
-      {
-        $configs=Config::find($id);
-        $configs->id=$request->id;
-        $configs->configkey=request()->configkey;
-        $configs->name=request()->name;
-        $configs->description=request()->description;
-        $configs->isActive=request()->isActive;
-        $configs->remark=request()->remark;
-        $configs->update();
-        return redirect('config/dashboard')
-        ->with('info','Successfully Update !');
-      }
      
     }
 
 
     public function delete($id)
     {
-      $data=Config::find($id);
-      $data->delete();
-      return back()->with('info',"Deleted !");
+
+      $response=Gate::inspect('only-admin');
+      if($response->allowed())
+      {
+        $data=Config::find($id);
+        $data->delete();
+        return back()->with('info',"Deleted !");
+      }
+      else
+      {
+        return back()->with('info',"You are not Authorized Delete");
+      }
+
+    
     }
 }
