+++ resources/views/staff/dashboard.blade.php
@@ -9,9 +9,14 @@
 		{{ session('info') }}
 	</div>
 	@endif
-  
+  @auth
+  @can('create-all')
   <a class="btn sm btn-primary" href="/staff/create" > +  Add New Staff</a><br><br>
-<table class="table table-bordered">
+  @endcan
+  @endauth
+
+  <h3>Staff Lists</h3>
+  <table class="table table-bordered">
 
   <thead>
    
@@ -21,6 +26,7 @@
       <th scope="col">Eamil</th>
       <th scope="col">Designation</th>
       <th scope="col">Phone</th>
+      <th scope="col">isActive</th>
       <th scope="col">remark</th>
       <th scope="col">Address</th>
       <th scope="col">Action</th>
@@ -38,12 +44,24 @@
       <td>{{$staffinfo->email}}</td>
       <td>{{$staffinfo->designation}}</td>
       <td>{{$staffinfo->phone}}</td>
+      <td>
+        @if ($staffinfo->isActive==0)
+        Inactive
+        @elseif($staffinfo->isActive==1)
+        Active
+        @else
+        Something Worng !
+        @endif
+      </td>
       <td>{{$staffinfo->remark}}</td>
       <td>{{$staffinfo->address}}</td>
      <td>
+      @can('update-all')
       <a class="btn sm btn-primary" href="/staff/detail/{{$staffinfo->id}}">Edit</a>
+      @endcan
+     @can('delete-all')
       <a class="btn sm btn-danger" href="/staff/delete/{{$staffinfo->id}}">Delete</a> 
-    
+    @endcan
     </td>
     
     </tr>
